Debug Branch for testing integration based on sample engine
Game system converted to use ISystem

Now able to send messages and target a specific entity's component

+++ TestFiles/Composition.h
@@ -4,7 +4,7 @@
 #include "IComponent.h"
 #include <vector>
 
-using EntityID = unsigned int;
+using EntityID = size_t;
 
 using ComponentArr = std::vector<Component*>;
